requirements.txt
selenium==4.5.0
python-environ==0.4.54
